refactor(storage): replace dynamic SQL with static queries (#9)
* refactor(storage): replace dynamic SQL with static queries in update_api_key

Replace dynamic SQL query construction with separate static queries for
each update combination. This improves type safety, maintainability, and
prevents potential SQL injection risks.

Changes:
- Replaced format!() string building with pattern matching
- Created 3 separate static SQL queries:
  * Update both is_active and description
  * Update only is_active
  * Update only description
- Maintained exact same functionality and API
- Preserved early return for no-op updates (None, None)

Benefits:
1. Type Safety: Static queries can be verified at compile time
2. Security: Eliminates string concatenation in SQL
3. Maintainability: Clear, explicit queries easier to understand
4. Performance: No runtime string allocation/formatting

Technical Details:
- Uses match expression on (is_active, &description) tuple
- Each branch has dedicated query with appropriate bindings
- Maintains consistent parameter ordering
- No behavioral changes - drop-in replacement

This is a security and code quality improvement with zero functional changes.

* refactor(storage): use sqlx::query! macro for compile-time query verification

Replace sqlx::query with sqlx::query! macro for all API key operations
to enable compile-time SQL verification and type checking. This provides
even stronger guarantees than the previous static query approach.

Changes:
- Updated all API key operations to use sqlx::query! macro:
  * create_api_key
  * get_api_key
  * get_api_key_by_hash
  * update_api_key_last_used
  * list_api_keys
  * update_api_key (all 3 branches)
  * delete_api_key

Benefits:
1. Compile-Time Verification: Queries are verified against the actual
   database schema at compile time
2. Type Safety: Return types are automatically inferred and checked
3. Column Validation: Column names and types are validated at compile time
4. SQL Syntax: Invalid SQL syntax caught during compilation, not runtime
5. Schema Changes: Compile errors if database schema changes
6. Zero Runtime Overhead: Same performance as regular queries

Technical Details:
- Uses sqlx-cli to generate query metadata at build time
- Queries verified against migrations in ./migrations directory
- Anonymous record types returned for SELECT queries
- Field access uses dot notation (r.field) instead of get("field")
- All existing functionality preserved with stronger guarantees

This enhancement builds on the previous static query refactor by adding
an additional layer of compile-time safety through sqlx's macro system.

Database compatibility verified through migrations.

* add sqlx offline mode, add db row models, use query_as, update some models' fields

---------

+++ src/config.rs
@@ -23,7 +23,7 @@ pub struct Config {
     pub master_api_key: String,
     pub default_session_ttl_seconds: u64,
     pub max_session_ttl_seconds: u64,
-    pub max_validation_attempts: i32,
+    pub max_validation_attempts: i64,
     pub cleanup_interval_seconds: u64,
     pub captcha_compression: u8,
 }
